pyproject.toml
[project]
name = "pyteletui"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "python-dotenv>=1.2.1",
    "python-telegram>=0.19.0",
    "textual>=6.11.0",
    "textual-dev>=1.8.0",
]
